Add files via upload
Herein is the final submission containing different plots cutting across line plot, bar plot, box plot, heat map and bubble plot. All were used to reveal the narrative of the adverse effect of climate change and the regions that will feel the most impact if remedial measures are not consistently initiated and protected.